Maria

+++ browser/views.py
@@ -16,32 +16,28 @@ from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.core.files.storage import FileSystemStorage
 from datetime import datetime
-import subprocess 
+import subprocess
 import uuid
 from django.http import JsonResponse, HttpResponseRedirect
 
 def check_job_status(request):
     job_id = request.GET.get('job_id')
-    #if not job_id:
-    #    return JsonResponse({'status': 'error', 'message': 'Job ID not provided.'}, status=400)
+    if not job_id:
+        return JsonResponse({'status': 'error', 'message': 'No job_id provided.'}, status=400)
 
-    job_status_path = os.path.join('data', 'jobs', job_id, 'status.txt')
+    job_status_path = os.path.join('/data/jobs', job_id, 'status.txt')
 
     if not os.path.exists(job_status_path):
-        return JsonResponse({'status': 'not_found'}, status=404)
+        return JsonResponse({'status': 'error', 'message': f'Status file not found at {job_status_path}'}, status=404)
 
-    try:
-        with open(job_status_path, 'r') as status_file:
-            status = status_file.read().strip()
-            if not status:
-                return JsonResponse({'status': 'in_progress'}, status=200)  
-    except Exception as e:
-        return JsonResponse({'status': 'error', 'message': str(e)}, status=500)
+    with open(job_status_path, 'r') as file:
+        job_status = file.read().strip()
 
-    if 'finished' in status:
-        return JsonResponse({'status': 'finished', 'redirect_url': '/results_job/?job_id=' + job_id}, status=200)
-
-    return JsonResponse({'status': status})
+    if job_status == 'finished':
+        redirect_url = reverse('results_job', args=[job_id])
+        return JsonResponse({'status': 'finished', 'redirect_url': redirect_url})
+    else:
+        return JsonResponse({'status': job_status, 'message': 'At the end of the process the results will be displayed.'})
 
 def contact_us(request):
     return render(request, 'browser/contact_us.html')
@@ -58,11 +54,20 @@ def search_view(request):
 def drosophiladb(request):
     return render(request, 'browser/drosophiladb.html')
 
-def job_running(request):
-    return render(request, 'browser/job_running.html')
+def job_running(request,job_id): 
+    job_status_path = os.path.join('/data/jobs/', job_id, 'status.txt')
 
+    if not os.path.exists(job_status_path):
+        return render(request, 'browser/error.html', {'message': 'Status file not found for this job.'}, status=404)
 
-#DATA = 'browser/static/jobs/Drosophila_ProteoCast/' #'/data/Drosophila_ProteoCast/'
+    try:
+        with open(job_status_path, 'r') as status_file:
+            status = status_file.read().strip()
+    except Exception as e:
+        return render(request, 'browser/error.html', {'message': f'Error reading status file: {str(e)}'}, status=500)
+
+    return render(request, 'browser/job_running.html', {'job_id': job_id, 'status': status})
+
 DATA = '/data/Drosophila_ProteoCast/'
 
 @csrf_exempt
@@ -76,6 +81,7 @@ def upload_file(request):
 
 def handle_upload(request, uploaded_file):
     now = datetime.now()
+    global job_id
     job_id = now.strftime('%Y-%m-%d_%H-%M-%S')
     folder_path = os.path.join('/data/jobs/', job_id)
     os.makedirs(folder_path, mode=0o755, exist_ok=True)
@@ -91,12 +97,13 @@ def handle_upload(request, uploaded_file):
             first_line = file.readline().strip()
         prot_name = first_line.lstrip('>')
         new_folder_path = '/data/jobs/' + prot_name
-        os.rename(folder_path, new_folder_path)
-        os.chdir(new_folder_path)
-        job_id = prot_name  # Actualiza el job_id con el nuevo nombre
+        job_id = prot_name
+        if not os.path.exists(new_folder_path):
+          os.rename(folder_path, new_folder_path)
+          os.chdir(new_folder_path)
 
-        run_docker_script = os.path.join(new_folder_path, 'run_docker.sh')
-        with open(run_docker_script, 'w') as script:
+          run_docker_script = os.path.join(new_folder_path, 'run_docker.sh')
+          with open(run_docker_script, 'w') as script:
             script.write(f"""#!/bin/bash
 #SBATCH --nodes=1
 #SBATCH --ntasks-per-node=1
@@ -109,18 +116,19 @@ def handle_upload(request, uploaded_file):
 
 docker run --rm -v "/data/jobs/{prot_name}:/opt/job" elodielaine/gemme:gemme /bin/bash -c "cd / && bash run.sh {uploaded_file.name}"
 """)
-        os.chmod(run_docker_script, 0o755)
+          os.chmod(run_docker_script, 0o755)
+  
+          subprocess.run(['sbatch', run_docker_script], check=True)
 
-        subprocess.run(['sbatch', run_docker_script], check=True)
-
-        job_status_path = os.path.join(new_folder_path, 'status.txt')
-        with open(job_status_path, 'w') as status_file:
+          job_status_path = os.path.join(new_folder_path, 'status.txt')
+          with open(job_status_path, 'w') as status_file:
             status_file.write('in_progress')
-
-        return JsonResponse({'redirect_url': f'/job_running/{job_id}'})
-
+        return JsonResponse({
+                'status': 'in_progress',
+                'job_id': job_id,
+                'redirect_url': f'/results/job_running/{job_id}'})
     except Exception as e:
-        return JsonResponse({'error': str(e)}, status=500)
+        return JsonResponse({'status': 'error', 'message': str(e)}, status=500)
 
 
 def serve_file(request, folder, filename):
@@ -131,29 +139,38 @@ def serve_file(request, folder, filename):
     else:
         return HttpResponse(f"File not found: {filename}", status=404)
 
-
-def results_job(request):
-    job_id = request.GET.get('job_id')
-    DATA = '/data/jobs/'+job_id
+def results_job(request,job_id):
+  try:
+    DATA_job = '/data/jobs/'+job_id
     alph = ["a","c","d","e","f","g","h","i","k","l","m","n","p","q","r","s","t","v","w","y"][::-1]
     alph = [i.upper() for i in alph]
 
-    ### Confidence values
-    proteocast_path = f'{DATA}{id_folder}/4.{FBpp_id}_ProteoCast.csv'
+    proteocast_path = f'/{DATA_job}/4.{job_id}_ProteoCast.csv'
     if not os.path.exists(proteocast_path):
-        return HttpResponse("ProteoCast file not found.")
-
-
+        return HttpResponse("ProteoCast file not found."+proteocast_path)
     df_proteocast = pd.read_csv(proteocast_path)
-    df_proteocast['GEMME_LocalConfidence'] = df_proteocast['GEMME_LocalConfidence'].replace({True:1, False:0})
+    def assign_variant_class(row):
+        if row['GEMME_score'] < -4:
+            return 'impactful'
+        elif -4 <= row['GEMME_score'] < -2:
+            return 'uncertain'
+        else:
+            return 'neutral'
+
+    df_proteocast['Variant_class'] = df_proteocast.apply(assign_variant_class, axis=1)
+    df_proteocast['GEMME_LocalConfidence'] = df_proteocast['GEMME_LocalConfidence'].replace({True: 1, False: 0})
     confidence_values = np.array(df_proteocast.groupby('Residue')['GEMME_LocalConfidence'].apply(lambda x: x.iloc[0]).tolist()).reshape(1, -1)
-
-
     df = pd.DataFrame(np.array(df_proteocast['GEMME_score']).reshape(20, -1, order='F'))
+    df_classes = pd.DataFrame(np.array(df_proteocast['Variant_class'].replace({'neutral': 1, 'uncertain': 2, 'impactful': 3})).reshape(20, -1, order='F'))
+    df_classesStr = pd.DataFrame(np.array(df_proteocast['Variant_class']).reshape(20, -1, order='F'))
     df_mut = pd.DataFrame(np.array(df_proteocast['Mutation']).reshape(20, -1, order='F'))
 
-
-
+    # Custom color scale for confidence
+    variantClasses_colorscale = [
+        [0, '#3688ED'],  # neutral- white
+        [0.5, '#E097CE'],  # uncertain - purple
+        [1, '#F25064']      #impactful - red
+    ]
     # Custom color scale for confidence
     confidence_colorscale = [
         [0, 'white'],  # Low confidence - white
@@ -167,77 +184,104 @@ def results_job(request):
         vertical_spacing=0.02,
     )
 
+    fig_VariantClasses = make_subplots(
+        rows=2, cols=1,
+        shared_xaxes=True,
+        row_heights=[0.9, 0.1],
+        vertical_spacing=0.02,
+    )
     heatmap_main = go.Heatmap(
         z=df.values[::-1],  # Reverse the order of rows
-        x=list(range(1, df.shape[1] + 1)),
+        x=list(range(1, df.shape[1])),
         y=alph,  # Reverse the order of y-axis labels
         colorscale=px.colors.sequential.Oranges[::-1],
+        showscale=False,
         customdata=df_mut.values[::-1],  # Reverse the order of custom data
         hovertemplate=(
             "Position: %{customdata}<br>"
             "Score: %{z:.2f}<extra></extra>"
         ),
     )
+    heatmap_classes = go.Heatmap(
+        z=df_classes.values[::-1],
+        x=list(range(1, df_classes.shape[1])),
+        y=alph,
+        customdata=np.dstack([df_mut.values[::-1], df_classesStr.values[::-1]]),
+        colorscale=variantClasses_colorscale,
+        showscale=False,
+        hovertemplate=(
+            "Position: %{customdata[0]}<br>"
+            "Class: %{customdata[1]}<extra></extra>"
+        ),
+        # These just control the gap size; their color is determined by the figure background.
+        xgap=0.3,
+        ygap=0.3,
+    )
 
     fig.add_trace(heatmap_main, row=1, col=1)
+    fig_VariantClasses.add_trace(heatmap_classes, row=1, col=1)
 
     heatmap_confidence = go.Heatmap(
        z=confidence_values,
+       x=list(range(1, df_classes.shape[1])),
        colorscale=confidence_colorscale,
        showscale=False,
-       hoverinfo='skip',
+       hovertemplate=(
+            "%{z}<extra></extra>"
+        ),
     )
+
     fig.add_trace(heatmap_confidence, row=2, col=1)
+    fig_VariantClasses.add_trace(heatmap_confidence, row=2, col=1)
 
     scatter_border = go.Scatter(
-       x=[0, df.shape[1], df.shape[1], 0, 0],
+       x=[0, df.shape[1]+1, df.shape[1]+1, 0, 0],
        y=[-0.5, -0.5, 0.5, 0.5, -0.5],
        mode="lines",
-       line=dict(color="black", width=2),
+       line=dict(color="darkblue", width=2),
        hoverinfo="skip",
-       showlegend=False
+       showlegend=False,
     )
     fig.add_trace(scatter_border, row=2, col=1)
+    fig_VariantClasses.add_trace(scatter_border, row=2, col=1)
+    fig.update_layout(title_x=1, autosize=False, width=1500, height=600, xaxis=dict(tickmode="array", tickvals=list(range(1, df.shape[1], 10)), ticktext=[str(i) for i in range(1, df.shape[1], 10)]), yaxis=dict(title="Substituting amino acid"), xaxis2=dict(title="Residue"))
+    fig_VariantClasses.update_layout(title_x=1, autosize=False, width=1500, height=600, xaxis=dict(tickmode="array", tickvals=list(range(1, df.shape[1]+1, 10)), ticktext=[str(i) for i in range(1, df.shape[1], 10)]), yaxis=dict(title="Substituting amino acid"),  xaxis2=dict(title="Residue"))
 
-    fig.update_layout(
-        title_x=1,
-        autosize=False,
-        width=1500,
-        height=600,
-        xaxis=dict(
-            tickmode="array",
-            tickvals=list(range(0, df.shape[1]+1, 10)),
-            ticktext=[str(i) for i in range(0,df.shape[1]+1, 10)],
-        ),
-        yaxis=dict(title="Mutations"),
-        coloraxis_colorbar=dict(
-            title="GEMME Score",
-            tickvals=[-8, -4, 0],
-        ),
-    )
 
     fig.update_yaxes(visible=False, row=2, col=1)
+    fig_VariantClasses.update_yaxes(visible=False, row=2, col=1)
     heatmap_html = fig.to_html(full_html=False)
+    heatmapClasses_html = fig_VariantClasses.to_html(full_html=False)
 
-    image_url_1 = f'/data/{id_folder}/6.{FBpp_id}_GMM.jpg'
-    pdb_url_1 = f'/data/{id_folder}/AF-Q45VV3-F1-model_v4.pdb'
-    fig_msarep = f'/data/{id_folder}/3.{FBpp_id}_msaRepresentation.jpg'
-    fig_segmentation = f'/data/{id_folder}/9.{FBpp_id}_SegProfile.png'
+    image_url_1 = f'/jobs/{job_id}/6.{job_id}_GMM.jpg'
+    pdb_url_1 = f'/data/2/AF-Q9W5X6-F1-model_v4.pdb'
+    fig_msarep = f'/jobs/{job_id}/3.{job_id}_msaRepresentation.jpg'
+    fig_segmentation = f'/jobs/{job_id}/9.{job_id}_SegProfile.png'
 
-    # Validate existence of files
-    for file_path in [image_url_1, pdb_url_1, fig_msarep, fig_segmentation]:
-        if not os.path.exists(file_path.replace('/data/', DATA)):
-            return HttpResponse(f"File not found: {file_path}")
+    files_exist = {
+        'image_url_1': os.path.exists(image_url_1),
+        'pdb_url_1': os.path.exists(pdb_url_1),
+        'fig_msarep': os.path.exists(fig_msarep),
+        'fig_segmentation': os.path.exists(fig_segmentation)
+    }
+
+    if not all(files_exist.values()):
+       missing_files = [k for k, v in files_exist.items() if not v]
+       error_message = f"Missing files: {', '.join(missing_files)}"
+#       return HttpResponse(f"Error: Missing files: {', '.join(missing_files)}", status=404)
 
     return render(request, 'browser/results.html', {
         'heatmap_html': heatmap_html,
-        'query': id_folder,
-        'prot_name': FBpp_id,
+        'heatmapClasses_html': heatmap_classes,
+        'query': job_id,
+        'prot_name': job_id,
         'image_url_1': image_url_1,
         'pdb_url_1': pdb_url_1,
         'fig_msarep': fig_msarep,
         'fig_segmentation': fig_segmentation,
     })
+  except Exception as e:
+        return HttpResponse(f"An error occurred: {str(e)}", status=500)
 
 
 def results_view(request):
